disk: move Partition and methods to separate file
Split up disk.go into smaller more manageable pieces: finish
by moving `Partition` to its own file.
